[DIDA] refactor. 코드정리 > userId도 Long타입으로 통일

+++ common/src/main/java/com/dida/common/actionhandler/CommentActionHandler.kt
@@ -2,5 +2,5 @@ package com.dida.common.actionhandler
 
 interface CommentActionHandler {
     fun onCommentMoreClicked(commentId: Long)
-    fun onCommentUserProfileClicked(userId: Int)
+    fun onCommentUserProfileClicked(userId: Long)
 }
