+++ data/src/main/java/com/dida/data/api/MainAPIService.kt
@@ -183,11 +183,11 @@ interface MainAPIService {
 
     // 타 유저 프로필 조회
     @GET("/user/{userId}")
-    suspend fun getUserUserId(@Path("userId") userId: Int): GetUserUserIdResponse
+    suspend fun getUserUserId(@Path("userId") userId: Long): GetUserUserIdResponse
 
     // 타 유저 카드 조회
     @GET("/user/cards/{userId}")
-    suspend fun getUserCardsUserId(@Path("userId") userId: Int): List<RecentCardResponse>
+    suspend fun getUserCardsUserId(@Path("userId") userId: Long): List<RecentCardResponse>
 
     @PATCH("/card/status/{cardId}")
     suspend fun patchDeleteNft(@Path("cardId") cardId: Long, @Body payPwd : String): Long
