+++ feature/home/src/main/java/com/dida/home/HomeActionHandler.kt
@@ -2,11 +2,11 @@ package com.dida.home
 
 interface HomeActionHandler {
     fun onHotItemClicked(cardId: Long)
-    fun onHotSellerItemClicked(userId: Int)
+    fun onHotSellerItemClicked(userId: Long)
     fun onSoldOutItemClicked(cardId: Long)
-    fun onCollectionItemClicked(userId: Int)
+    fun onCollectionItemClicked(userId: Long)
     fun onSoldOutDayClicked(day: Int)
-    fun onUserFollowClicked(userId: Int)
+    fun onUserFollowClicked(userId: Long)
     fun onHotSellerMoreClicked()
     fun onSoldOutMoreClicked()
     fun onRecentMoreNftClicked()
