+++ feature/hot-user/src/main/java/com/dida/hot_user/HotUserNavigationAction.kt
@@ -1,6 +1,6 @@
 package com.dida.hot_user
 
 sealed class HotUserNavigationAction {
-    class NavigateToUserProfile(val userId: Int): HotUserNavigationAction()
+    class NavigateToUserProfile(val userId: Long): HotUserNavigationAction()
     object NavigateToFollow: HotUserNavigationAction()
 }
