+++ model/src/main/java/com/dida/model/data/response/GetCardResponse.kt
@@ -7,7 +7,7 @@ data class GetCardResponse(
     @SerializedName("title") val title: String,
     @SerializedName("description") val description: String,
     @SerializedName("imgUrl") val imgUrl: String,
-    @SerializedName("userId") val userId: Int,
+    @SerializedName("userId") val userId: Long,
     @SerializedName("nickname") val nickname: String,
     @SerializedName("profileUrl") val profileUrl: String,
     @SerializedName("id") val id: String,
